requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-cors-headers==3.7.0
django-excel==0.0.10
django-import-export==2.5.0
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
lml==0.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
phonenumbers==8.12.17
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
tablib==3.0.0
texttable==1.6.3
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
